services/usage-stats/requirements.txt
requests
pymysql
python-dotenv
sqlalchemy
psycopg2
asyncpg
asyncmy